refactor(mcp): switch services to cubepi_admin_discovery/refresh
